requirements.txt
boto3
aws-lambda-powertools
pydantic
pyJWT == 2.8.0